Create QuickBogo.py
Works like quick sort, kind of, except worse in every way imaginable.

Randomly chooses a pivot, and then randomly sorts elements from the array into two arrays. If everything in the first array is less than or equal to the pivot, and everything in the second array is greater or equal to than the pivot, then the iteration is considered successful. If not, a new pivot is chosen and the process repeated.

On successful iteration, like quick sort, these two arrays have the above process repeated on them. This is recursively implemented.
Each iteration has a (1/2)^(n-1) chance of being correctly split, meaning that on average the first iteration has an expected number of iterations of O(2^(n-1)). On average, the successive arrays will be halved in size on each iteration, and since 2^(n-1) >> 2^0.5(n-1) this can be considered negligible. Assuming absolute best case scenario, the performance would be O(log(n)). Naturally, worst case performance is to never finish at all O(inf).